refactor: rename misc folder to utils and convert it to Python package (#40)
