+++ sftool/utils/errors.py
