+++ sftool/utils/runtime.py
@@ -6,7 +6,7 @@ required external tools and resources exist before execution.
 """
 
 import os
-from sftool.misc.errors import RuntimeDependencyError
+from sftool.utils.errors import RuntimeDependencyError
 from sftool.core.config import PathsConfig
