+++ sftool/utils/sfdiagram.png
