+++ sftool/utils/utils.py
